@@ -7,42 +7,48 @@ import os
 import re
 import sys
 import traceback
+import warnings
 import zipfile
+from collections.abc import Awaitable, Callable
 from functools import lru_cache
 from io import BytesIO, TextIOWrapper
 from pathlib import Path
 from tempfile import NamedTemporaryFile, TemporaryFile
-from typing import Dict, List, Optional
+from typing import Annotated, Any, Optional, TypeVar
 
 import soundfile
 import uvicorn
-from fastapi import FastAPI, Form, HTTPException, Query, Request, Response
+from fastapi import Depends, FastAPI, Form, HTTPException, Query, Request, Response
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.openapi.utils import get_openapi
-from fastapi.responses import HTMLResponse, JSONResponse
+from fastapi.responses import JSONResponse
 from fastapi.templating import Jinja2Templates
-from pydantic import ValidationError, conint
+from pydantic import ValidationError
 from starlette.background import BackgroundTask
+from starlette.middleware.errors import ServerErrorMiddleware
 from starlette.responses import FileResponse
 
 from voicevox_engine import __version__
 from voicevox_engine.cancellable_engine import CancellableEngine
-from voicevox_engine.downloadable_library import LibraryManager
+from voicevox_engine.core_adapter import CoreAdapter
+from voicevox_engine.core_initializer import initialize_cores
 from voicevox_engine.engine_manifest import EngineManifestLoader
 from voicevox_engine.engine_manifest.EngineManifest import EngineManifest
-from voicevox_engine.kana_parser import create_kana, parse_kana
+from voicevox_engine.library_manager import LibraryManager
+from voicevox_engine.metas.Metas import StyleId
 from voicevox_engine.metas.MetasStore import MetasStore, construct_lookup
 from voicevox_engine.model import (
     AccentPhrase,
     AudioQuery,
-    DownloadableLibrary,
-    InstalledLibrary,
+    BaseLibraryInfo,
+    DownloadableLibraryInfo,
+    InstalledLibraryInfo,
     MorphableTargetInfo,
     ParseKanaBadRequest,
     ParseKanaError,
     Speaker,
     SpeakerInfo,
-    SpeakerNotFoundError,
+    StyleIdNotFoundError,
     SupportedDevicesInfo,
     UserDictWord,
     VvlibManifest,
@@ -64,7 +70,11 @@ from voicevox_engine.setting import (
     Setting,
     SettingLoader,
 )
-from voicevox_engine.synthesis_engine import SynthesisEngineBase, make_synthesis_engines
+from voicevox_engine.tts_pipeline.kana_converter import create_kana, parse_kana
+from voicevox_engine.tts_pipeline.tts_engine import (
+    TTSEngine,
+    make_tts_engines_from_cores,
+)
 from voicevox_engine.user_dict import (
     apply_word,
     delete_word,
@@ -80,7 +90,27 @@ from voicevox_engine.utility import (
     engine_root,
     get_latest_core_version,
     get_save_dir,
+    internal_root,
 )
+from voicevox_engine.utility.run_utility import decide_boolean_from_env
+
+# NOTE: Python 3.12以降で[S: StyleId | list[StyleId]]に置き換えられる
+S = TypeVar("S", StyleId, list[StyleId])
+
+
+def get_style_id_from_deprecated(style_id: S | None, deprecated_speaker: S | None) -> S:
+    """
+    style_idとspeaker両方ともNoneかNoneでないかをチェックし、
+    どちらか片方しかNoneが存在しなければstyle_idを返す
+    """
+    if deprecated_speaker is not None and style_id is None:
+        warnings.warn("speakerは非推奨です。style_idを利用してください。", stacklevel=1)
+        return deprecated_speaker
+    elif style_id is not None and deprecated_speaker is None:
+        return style_id
+    raise HTTPException(
+        status_code=400, detail="speakerとstyle_idが両方とも存在しないか、両方とも存在しています。"
+    )
 
 
 def b64encode_str(s):
@@ -93,43 +123,62 @@ def set_output_log_utf8() -> None:
     """
     # コンソールがない環境だとNone https://docs.python.org/ja/3/library/sys.html#sys.__stdin__
     if sys.stdout is not None:
-        # 必ずしもreconfigure()が実装されているとは限らない
-        try:
+        if isinstance(sys.stdout, TextIOWrapper):
             sys.stdout.reconfigure(encoding="utf-8")
-        except AttributeError:
+        else:
             # バッファを全て出力する
             sys.stdout.flush()
-            sys.stdout = TextIOWrapper(
-                sys.stdout.buffer, encoding="utf-8", errors="backslashreplace"
-            )
+            try:
+                sys.stdout = TextIOWrapper(
+                    sys.stdout.buffer, encoding="utf-8", errors="backslashreplace"
+                )
+            except AttributeError:
+                # stdout.bufferがない場合は無視
+                pass
     if sys.stderr is not None:
-        try:
+        if isinstance(sys.stderr, TextIOWrapper):
             sys.stderr.reconfigure(encoding="utf-8")
-        except AttributeError:
+        else:
             sys.stderr.flush()
-            sys.stderr = TextIOWrapper(
-                sys.stderr.buffer, encoding="utf-8", errors="backslashreplace"
-            )
+            try:
+                sys.stderr = TextIOWrapper(
+                    sys.stderr.buffer, encoding="utf-8", errors="backslashreplace"
+                )
+            except AttributeError:
+                # stderr.bufferがない場合は無視
+                pass
 
 
 def generate_app(
-    synthesis_engines: Dict[str, SynthesisEngineBase],
+    tts_engines: dict[str, TTSEngine],
+    cores: dict[str, CoreAdapter],
     latest_core_version: str,
     setting_loader: SettingLoader,
+    preset_manager: PresetManager,
+    cancellable_engine: CancellableEngine | None = None,
     root_dir: Optional[Path] = None,
     cors_policy_mode: CorsPolicyMode = CorsPolicyMode.localapps,
-    allow_origin: Optional[List[str]] = None,
+    allow_origin: Optional[list[str]] = None,
+    disable_mutable_api: bool = False,
 ) -> FastAPI:
     if root_dir is None:
         root_dir = engine_root()
-
-    default_sampling_rate = synthesis_engines[latest_core_version].default_sampling_rate
 
     app = FastAPI(
         title="VOICEVOX Engine",
         description="VOICEVOXの音声合成エンジンです。",
         version=__version__,
     )
+
+    # 未処理の例外が発生するとCORSMiddlewareが適用されない問題に対するワークアラウンド
+    # ref: https://github.com/VOICEVOX/voicevox_engine/issues/91
+    async def global_execution_handler(request: Request, exc: Exception) -> Response:
+        return JSONResponse(
+            status_code=500,
+            content="Internal Server Error",
+        )
+
+    app.add_middleware(ServerErrorMiddleware, handler=global_execution_handler)
 
     # CORS用のヘッダを生成するミドルウェア
     localhost_regex = "^https?://(localhost|127\\.0\\.0\\.1)(:[0-9]+)?$"
@@ -157,7 +206,9 @@ def generate_app(
 
     # 許可されていないOriginを遮断するミドルウェア
     @app.middleware("http")
-    async def block_origin_middleware(request: Request, call_next):
+    async def block_origin_middleware(
+        request: Request, call_next: Callable[[Request], Awaitable[Response]]
+    ) -> Response | JSONResponse:
         isValidOrigin: bool = False
         if "Origin" not in request.headers:  # Originのない純粋なリクエストの場合
             isValidOrigin = True
@@ -177,9 +228,11 @@ def generate_app(
                 status_code=403, content={"detail": "Origin not allowed"}
             )
 
-    preset_manager = PresetManager(
-        preset_path=root_dir / "presets.yaml",
-    )
+    # 許可されていないAPIを無効化する
+    def check_disabled_mutable_api():
+        if disable_mutable_api:
+            raise HTTPException(status_code=403, detail="エンジンの静的なデータを変更するAPIは無効化されています")
+
     engine_manifest_data = EngineManifestLoader(
         engine_root() / "engine_manifest.json", engine_root()
     ).load_manifest()
@@ -193,7 +246,11 @@ def generate_app(
 
     metas_store = MetasStore(root_dir / "speaker_info")
 
-    setting_ui_template = Jinja2Templates(directory=engine_root() / "ui_template")
+    setting_ui_template = Jinja2Templates(
+        directory=internal_root() / "ui_template",
+        variable_start_string="<JINJA_PRE>",
+        variable_end_string="<JINJA_POST>",
+    )
 
     # キャッシュを有効化
     # モジュール側でlru_cacheを指定するとキャッシュを制御しにくいため、HTTPサーバ側で指定する
@@ -202,7 +259,7 @@ def generate_app(
 
     # @app.on_event("startup")
     # async def start_catch_disconnection():
-    #     if args.enable_cancellable_synthesis:
+    #     if cancellable_engine is not None:
     #         loop = asyncio.get_event_loop()
     #         _ = loop.create_task(cancellable_engine.catch_disconnection())
 
@@ -210,11 +267,19 @@ def generate_app(
     def apply_user_dict():
         update_dict()
 
-    def get_engine(core_version: Optional[str]) -> SynthesisEngineBase:
+    def get_engine(core_version: Optional[str]) -> TTSEngine:
         if core_version is None:
-            return synthesis_engines[latest_core_version]
-        if core_version in synthesis_engines:
-            return synthesis_engines[core_version]
+            return tts_engines[latest_core_version]
+        if core_version in tts_engines:
+            return tts_engines[core_version]
+        raise HTTPException(status_code=422, detail="不明なバージョンです")
+
+    def get_core(core_version: Optional[str]) -> CoreAdapter:
+        """指定したバージョンのコアを取得する"""
+        if core_version is None:
+            return cores[latest_core_version]
+        if core_version in cores:
+            return cores[core_version]
         raise HTTPException(status_code=422, detail="不明なバージョンです")
 
     @app.post(
@@ -223,12 +288,21 @@ def generate_app(
         tags=["クエリ作成"],
         summary="音声合成用のクエリを作成する",
     )
-    def audio_query(text: str, speaker: int, core_version: Optional[str] = None):
+    def audio_query(
+        text: str,
+        style_id: StyleId | None = Query(default=None),  # noqa: B008
+        speaker: StyleId | None = Query(default=None, deprecated=True),  # noqa: B008
+        core_version: str | None = None,
+    ) -> AudioQuery:
         """
-        クエリの初期値を得ます。ここで得られたクエリはそのまま音声合成に利用できます。各値の意味は`Schemas`を参照してください。
+        音声合成用のクエリの初期値を得ます。ここで得られたクエリはそのまま音声合成に利用できます。各値の意味は`Schemas`を参照してください。
         """
+        style_id = get_style_id_from_deprecated(
+            style_id=style_id, deprecated_speaker=speaker
+        )
         engine = get_engine(core_version)
-        accent_phrases = engine.create_accent_phrases(text, speaker_id=speaker)
+        core = get_core(core_version)
+        accent_phrases = engine.create_accent_phrases(text, style_id)
         return AudioQuery(
             accent_phrases=accent_phrases,
             speedScale=1,
@@ -237,7 +311,7 @@ def generate_app(
             volumeScale=1,
             prePhonemeLength=0.1,
             postPhonemeLength=0.1,
-            outputSamplingRate=default_sampling_rate,
+            outputSamplingRate=core.default_sampling_rate,
             outputStereo=False,
             kana=create_kana(accent_phrases),
         )
@@ -249,12 +323,15 @@ def generate_app(
         summary="音声合成用のクエリをプリセットを用いて作成する",
     )
     def audio_query_from_preset(
-        text: str, preset_id: int, core_version: Optional[str] = None
-    ):
+        text: str,
+        preset_id: int,
+        core_version: str | None = None,
+    ) -> AudioQuery:
         """
-        クエリの初期値を得ます。ここで得られたクエリはそのまま音声合成に利用できます。各値の意味は`Schemas`を参照してください。
+        音声合成用のクエリの初期値を得ます。ここで得られたクエリはそのまま音声合成に利用できます。各値の意味は`Schemas`を参照してください。
         """
         engine = get_engine(core_version)
+        core = get_core(core_version)
         try:
             presets = preset_manager.load_presets()
         except PresetError as err:
@@ -266,9 +343,7 @@ def generate_app(
         else:
             raise HTTPException(status_code=422, detail="該当するプリセットIDが見つかりません")
 
-        accent_phrases = engine.create_accent_phrases(
-            text, speaker_id=selected_preset.style_id
-        )
+        accent_phrases = engine.create_accent_phrases(text, selected_preset.style_id)
         return AudioQuery(
             accent_phrases=accent_phrases,
             speedScale=selected_preset.speedScale,
@@ -277,14 +352,14 @@ def generate_app(
             volumeScale=selected_preset.volumeScale,
             prePhonemeLength=selected_preset.prePhonemeLength,
             postPhonemeLength=selected_preset.postPhonemeLength,
-            outputSamplingRate=default_sampling_rate,
+            outputSamplingRate=core.default_sampling_rate,
             outputStereo=False,
             kana=create_kana(accent_phrases),
         )
 
     @app.post(
         "/accent_phrases",
-        response_model=List[AccentPhrase],
+        response_model=list[AccentPhrase],
         tags=["クエリ編集"],
         summary="テキストからアクセント句を得る",
         responses={
@@ -296,81 +371,87 @@ def generate_app(
     )
     def accent_phrases(
         text: str,
-        speaker: int,
+        style_id: StyleId | None = Query(default=None),  # noqa: B008
+        speaker: StyleId | None = Query(default=None, deprecated=True),  # noqa: B008
         is_kana: bool = False,
-        core_version: Optional[str] = None,
-    ):
+        core_version: str | None = None,
+    ) -> list[AccentPhrase]:
         """
         テキストからアクセント句を得ます。
-        is_kanaが`true`のとき、テキストは次のようなAquesTalkライクな記法に従う読み仮名として処理されます。デフォルトは`false`です。
+        is_kanaが`true`のとき、テキストは次のAquesTalk 風記法で解釈されます。デフォルトは`false`です。
         * 全てのカナはカタカナで記述される
         * アクセント句は`/`または`、`で区切る。`、`で区切った場合に限り無音区間が挿入される。
         * カナの手前に`_`を入れるとそのカナは無声化される
         * アクセント位置を`'`で指定する。全てのアクセント句にはアクセント位置を1つ指定する必要がある。
         * アクセント句末に`？`(全角)を入れることにより疑問文の発音ができる。
         """
+        style_id = get_style_id_from_deprecated(
+            style_id=style_id, deprecated_speaker=speaker
+        )
         engine = get_engine(core_version)
         if is_kana:
             try:
-                accent_phrases = parse_kana(text)
+                return engine.create_accent_phrases_from_kana(text, style_id)
             except ParseKanaError as err:
                 raise HTTPException(
-                    status_code=400,
-                    detail=ParseKanaBadRequest(err).dict(),
+                    status_code=400, detail=ParseKanaBadRequest(err).dict()
                 )
-            accent_phrases = engine.replace_mora_data(
-                accent_phrases=accent_phrases, speaker_id=speaker
-            )
-
-            return accent_phrases
         else:
-            return engine.create_accent_phrases(text, speaker_id=speaker)
+            return engine.create_accent_phrases(text, style_id)
 
     @app.post(
         "/mora_data",
-        response_model=List[AccentPhrase],
+        response_model=list[AccentPhrase],
         tags=["クエリ編集"],
         summary="アクセント句から音高・音素長を得る",
     )
     def mora_data(
-        accent_phrases: List[AccentPhrase],
-        speaker: int,
-        core_version: Optional[str] = None,
-    ):
+        accent_phrases: list[AccentPhrase],
+        style_id: StyleId | None = Query(default=None),  # noqa: B008
+        speaker: StyleId | None = Query(default=None, deprecated=True),  # noqa: B008
+        core_version: str | None = None,
+    ) -> list[AccentPhrase]:
+        style_id = get_style_id_from_deprecated(
+            style_id=style_id, deprecated_speaker=speaker
+        )
         engine = get_engine(core_version)
-        return engine.replace_mora_data(accent_phrases, speaker_id=speaker)
+        return engine.update_length_and_pitch(accent_phrases, style_id)
 
     @app.post(
         "/mora_length",
-        response_model=List[AccentPhrase],
+        response_model=list[AccentPhrase],
         tags=["クエリ編集"],
         summary="アクセント句から音素長を得る",
     )
     def mora_length(
-        accent_phrases: List[AccentPhrase],
-        speaker: int,
-        core_version: Optional[str] = None,
-    ):
-        engine = get_engine(core_version)
-        return engine.replace_phoneme_length(
-            accent_phrases=accent_phrases, speaker_id=speaker
+        accent_phrases: list[AccentPhrase],
+        style_id: StyleId | None = Query(default=None),  # noqa: B008
+        speaker: StyleId | None = Query(default=None, deprecated=True),  # noqa: B008
+        core_version: str | None = None,
+    ) -> list[AccentPhrase]:
+        style_id = get_style_id_from_deprecated(
+            style_id=style_id, deprecated_speaker=speaker
         )
+        engine = get_engine(core_version)
+        return engine.update_length(accent_phrases, style_id)
 
     @app.post(
         "/mora_pitch",
-        response_model=List[AccentPhrase],
+        response_model=list[AccentPhrase],
         tags=["クエリ編集"],
         summary="アクセント句から音高を得る",
     )
     def mora_pitch(
-        accent_phrases: List[AccentPhrase],
-        speaker: int,
-        core_version: Optional[str] = None,
-    ):
-        engine = get_engine(core_version)
-        return engine.replace_mora_pitch(
-            accent_phrases=accent_phrases, speaker_id=speaker
+        accent_phrases: list[AccentPhrase],
+        style_id: StyleId | None = Query(default=None),  # noqa: B008
+        speaker: StyleId | None = Query(default=None, deprecated=True),  # noqa: B008
+        core_version: str | None = None,
+    ) -> list[AccentPhrase]:
+        style_id = get_style_id_from_deprecated(
+            style_id=style_id, deprecated_speaker=speaker
         )
+        engine = get_engine(core_version)
+        return engine.update_pitch(accent_phrases, style_id)
 
     @app.post(
         "/synthesis",
@@ -387,18 +468,20 @@ def generate_app(
     )
     def synthesis(
         query: AudioQuery,
-        speaker: int,
+        style_id: StyleId | None = Query(default=None),  # noqa: B008
+        speaker: StyleId | None = Query(default=None, deprecated=True),  # noqa: B008
         enable_interrogative_upspeak: bool = Query(  # noqa: B008
             default=True,
             description="疑問系のテキストが与えられたら語尾を自動調整する",
         ),
-        core_version: Optional[str] = None,
-    ):
+        core_version: str | None = None,
+    ) -> FileResponse:
+        style_id = get_style_id_from_deprecated(
+            style_id=style_id, deprecated_speaker=speaker
+        )
         engine = get_engine(core_version)
-        wave = engine.synthesis(
-            query=query,
-            speaker_id=speaker,
-            enable_interrogative_upspeak=enable_interrogative_upspeak,
+        wave = engine.synthesize_wave(
+            query, style_id, enable_interrogative_upspeak=enable_interrogative_upspeak
         )
 
         with NamedTemporaryFile(delete=False) as f:
@@ -427,20 +510,21 @@ def generate_app(
     )
     def cancellable_synthesis(
         query: AudioQuery,
-        speaker: int,
         request: Request,
-        core_version: Optional[str] = None,
-    ):
-        if not args.enable_cancellable_synthesis:
+        style_id: StyleId | None = Query(default=None),  # noqa: B008
+        speaker: StyleId | None = Query(default=None, deprecated=True),  # noqa: B008
+        core_version: str | None = None,
+    ) -> FileResponse:
+        style_id = get_style_id_from_deprecated(
+            style_id=style_id, deprecated_speaker=speaker
+        )
+        if cancellable_engine is None:
             raise HTTPException(
                 status_code=404,
                 detail="実験的機能はデフォルトで無効になっています。使用するには引数を指定してください。",
             )
         f_name = cancellable_engine._synthesis_impl(
-            query=query,
-            speaker_id=speaker,
-            request=request,
-            core_version=core_version,
+            query, style_id, request, core_version=core_version
         )
         if f_name == "":
             raise HTTPException(status_code=422, detail="不明なバージョンです")
@@ -467,27 +551,27 @@ def generate_app(
         summary="複数まとめて音声合成する",
     )
     def multi_synthesis(
-        queries: List[AudioQuery],
-        speaker: int,
-        core_version: Optional[str] = None,
-    ):
+        queries: list[AudioQuery],
+        style_id: StyleId | None = Query(default=None),  # noqa: B008
+        speaker: StyleId | None = Query(default=None, deprecated=True),  # noqa: B008
+        core_version: str | None = None,
+    ) -> FileResponse:
+        style_id = get_style_id_from_deprecated(
+            style_id=style_id, deprecated_speaker=speaker
+        )
         engine = get_engine(core_version)
         sampling_rate = queries[0].outputSamplingRate
 
         with NamedTemporaryFile(delete=False) as f:
-
             with zipfile.ZipFile(f, mode="a") as zip_file:
-
                 for i in range(len(queries)):
-
                     if queries[i].outputSamplingRate != sampling_rate:
                         raise HTTPException(
                             status_code=422, detail="サンプリングレートが異なるクエリがあります"
                         )
 
                     with TemporaryFile() as wav_file:
-
-                        wave = engine.synthesis(query=queries[i], speaker_id=speaker)
+                        wave = engine.synthesize_wave(queries[i], style_id)
                         soundfile.write(
                             file=wav_file,
                             data=wave,
@@ -505,35 +589,34 @@ def generate_app(
 
     @app.post(
         "/morphable_targets",
-        response_model=List[Dict[str, MorphableTargetInfo]],
+        response_model=list[dict[str, MorphableTargetInfo]],
         tags=["音声合成"],
-        summary="指定した話者に対してエンジン内の話者がモーフィングが可能か判定する",
+        summary="指定したスタイルに対してエンジン内の話者がモーフィングが可能か判定する",
     )
     def morphable_targets(
-        base_speakers: List[int],
-        core_version: Optional[str] = None,
-    ):
+        base_style_ids: list[StyleId], core_version: str | None = None
+    ) -> list[dict[str, MorphableTargetInfo]]:
         """
-        指定されたベース話者に対してエンジン内の各話者がモーフィング機能を利用可能か返します。
+        指定されたベーススタイルに対してエンジン内の各話者がモーフィング機能を利用可能か返します。
         モーフィングの許可/禁止は`/speakers`の`speaker.supported_features.synthesis_morphing`に記載されています。
         プロパティが存在しない場合は、モーフィングが許可されているとみなします。
         返り値の話者はstring型なので注意。
         """
-        engine = get_engine(core_version)
+        core = get_core(core_version)
 
         try:
-            speakers = metas_store.load_combined_metas(engine=engine)
+            speakers = metas_store.load_combined_metas(core=core)
             morphable_targets = get_morphable_targets(
-                speakers=speakers, base_speakers=base_speakers
+                speakers=speakers, base_style_ids=base_style_ids
             )
             # jsonはint型のキーを持てないので、string型に変換する
             return [
                 {str(k): v for k, v in morphable_target.items()}
                 for morphable_target in morphable_targets
             ]
-        except SpeakerNotFoundError as e:
+        except StyleIdNotFoundError as e:
             raise HTTPException(
-                status_code=404, detail=f"該当する話者(speaker={e.speaker})が見つかりません"
+                status_code=404, detail=f"該当するスタイル(style_id={e.style_id})が見つかりません"
             )
 
     @app.post(
@@ -547,43 +630,57 @@ def generate_app(
             }
         },
         tags=["音声合成"],
-        summary="2人の話者でモーフィングした音声を合成する",
+        summary="2種類のスタイルでモーフィングした音声を合成する",
     )
     def _synthesis_morphing(
         query: AudioQuery,
-        base_speaker: int,
-        target_speaker: int,
+        base_style_id: StyleId | None = Query(default=None),  # noqa: B008
+        base_speaker: (StyleId | None) = Query(  # noqa: B008
+            default=None, deprecated=True
+        ),
+        target_style_id: StyleId | None = Query(default=None),  # noqa: B008
+        target_speaker: (StyleId | None) = Query(  # noqa: B008
+            default=None, deprecated=True
+        ),
         morph_rate: float = Query(..., ge=0.0, le=1.0),  # noqa: B008
-        core_version: Optional[str] = None,
-    ):
+        core_version: str | None = None,
+    ) -> FileResponse:
         """
-        指定された2人の話者で音声を合成、指定した割合でモーフィングした音声を得ます。
-        モーフィングの割合は`morph_rate`で指定でき、0.0でベースの話者、1.0でターゲットの話者に近づきます。
+        指定された2種類のスタイルで音声を合成、指定した割合でモーフィングした音声を得ます。
+        モーフィングの割合は`morph_rate`で指定でき、0.0でベースのスタイル、1.0でターゲットのスタイルに近づきます。
         """
+        base_style_id = get_style_id_from_deprecated(
+            style_id=base_style_id, deprecated_speaker=base_speaker
+        )
+        target_style_id = get_style_id_from_deprecated(
+            style_id=target_style_id, deprecated_speaker=target_speaker
+        )
         engine = get_engine(core_version)
+        core = get_core(core_version)
 
         try:
-            speakers = metas_store.load_combined_metas(engine=engine)
+            speakers = metas_store.load_combined_metas(core=core)
             speaker_lookup = construct_lookup(speakers=speakers)
             is_permitted = is_synthesis_morphing_permitted(
-                speaker_lookup, base_speaker, target_speaker
+                speaker_lookup, base_style_id, target_style_id
             )
             if not is_permitted:
                 raise HTTPException(
                     status_code=400,
                     detail="指定された話者ペアでのモーフィングはできません",
                 )
-        except SpeakerNotFoundError as e:
+        except StyleIdNotFoundError as e:
             raise HTTPException(
-                status_code=404, detail=f"該当する話者(speaker={e.speaker})が見つかりません"
+                status_code=404, detail=f"該当するスタイル(style_id={e.style_id})が見つかりません"
             )
 
         # 生成したパラメータはキャッシュされる
         morph_param = synthesis_morphing_parameter(
             engine=engine,
+            core=core,
             query=query,
-            base_speaker=base_speaker,
-            target_speaker=target_speaker,
+            base_style_id=base_style_id,
+            target_style_id=target_style_id,
         )
 
         morph_wave = synthesis_morphing(
@@ -620,14 +717,14 @@ def generate_app(
         tags=["その他"],
         summary="base64エンコードされた複数のwavデータを一つに結合する",
     )
-    def connect_waves(waves: List[str]):
+    def connect_waves(waves: list[str]) -> FileResponse:
         """
         base64エンコードされたwavデータを一纏めにし、wavファイルで返します。
         """
         try:
             waves_nparray, sampling_rate = connect_base64_waves(waves)
         except ConnectBase64WavesException as err:
-            return HTTPException(status_code=422, detail=str(err))
+            raise HTTPException(status_code=422, detail=str(err))
 
         with NamedTemporaryFile(delete=False) as f:
             soundfile.write(
@@ -643,14 +740,14 @@ def generate_app(
             background=BackgroundTask(delete_file, f.name),
         )
 
-    @app.get("/presets", response_model=List[Preset], tags=["その他"])
-    def get_presets():
+    @app.get("/presets", response_model=list[Preset], tags=["その他"])
+    def get_presets() -> list[Preset]:
         """
         エンジンが保持しているプリセットの設定を返します
 
         Returns
         -------
-        presets: List[Preset]
+        presets: list[Preset]
             プリセットのリスト
         """
         try:
@@ -659,8 +756,13 @@ def generate_app(
             raise HTTPException(status_code=422, detail=str(err))
         return presets
 
-    @app.post("/add_preset", response_model=int, tags=["その他"])
-    def add_preset(preset: Preset):
+    @app.post(
+        "/add_preset",
+        response_model=int,
+        tags=["その他"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
+    def add_preset(preset: Preset) -> int:
         """
         新しいプリセットを追加します
 
@@ -681,8 +783,13 @@ def generate_app(
             raise HTTPException(status_code=422, detail=str(err))
         return id
 
-    @app.post("/update_preset", response_model=int, tags=["その他"])
-    def update_preset(preset: Preset):
+    @app.post(
+        "/update_preset",
+        response_model=int,
+        tags=["その他"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
+    def update_preset(preset: Preset) -> int:
         """
         既存のプリセットを更新します
 
@@ -703,8 +810,13 @@ def generate_app(
             raise HTTPException(status_code=422, detail=str(err))
         return id
 
-    @app.post("/delete_preset", status_code=204, tags=["その他"])
-    def delete_preset(id: int):
+    @app.post(
+        "/delete_preset",
+        status_code=204,
+        tags=["その他"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
+    def delete_preset(id: int) -> Response:
         """
         既存のプリセットを削除します
 
@@ -724,22 +836,24 @@ def generate_app(
     def version() -> str:
         return __version__
 
-    @app.get("/core_versions", response_model=List[str], tags=["その他"])
-    def core_versions() -> List[str]:
+    @app.get("/core_versions", response_model=list[str], tags=["その他"])
+    def core_versions() -> Response:
         return Response(
-            content=json.dumps(list(synthesis_engines.keys())),
+            content=json.dumps(list(cores.keys())),
             media_type="application/json",
         )
 
-    @app.get("/speakers", response_model=List[Speaker], tags=["その他"])
+    @app.get("/speakers", response_model=list[Speaker], tags=["その他"])
     def speakers(
-        core_version: Optional[str] = None,
-    ):
-        engine = get_engine(core_version)
-        return metas_store.load_combined_metas(engine=engine)
+        core_version: str | None = None,
+    ) -> list[Speaker]:
+        return metas_store.load_combined_metas(get_core(core_version))
 
     @app.get("/speaker_info", response_model=SpeakerInfo, tags=["その他"])
-    def speaker_info(speaker_uuid: str, core_version: Optional[str] = None):
+    def speaker_info(
+        speaker_uuid: str,
+        core_version: str | None = None,
+    ) -> dict[str, Any]:
         """
         指定されたspeaker_uuidに関する情報をjson形式で返します。
         画像や音声はbase64エンコードされたものが返されます。
@@ -748,7 +862,32 @@ def generate_app(
         -------
         ret_data: SpeakerInfo
         """
-        speakers = json.loads(get_engine(core_version).speakers)
+
+        # エンジンに含まれる話者メタ情報は、次のディレクトリ構造に従わなければならない：
+        # {root_dir}/
+        #   speaker_info/
+        #       {speaker_uuid_0}/
+        #           policy.md
+        #           portrait.png
+        #           icons/
+        #               {id_0}.png
+        #               {id_1}.png
+        #               ...
+        #           portraits/
+        #               {id_0}.png
+        #               {id_1}.png
+        #               ...
+        #           voice_samples/
+        #               {id_0}_001.wav
+        #               {id_0}_002.wav
+        #               {id_0}_003.wav
+        #               {id_1}_001.wav
+        #               ...
+        #       {speaker_uuid_1}/
+        #           ...
+
+        # 該当話者の検索
+        speakers = json.loads(get_core(core_version).speakers)
         for i in range(len(speakers)):
             if speakers[i]["speaker_uuid"] == speaker_uuid:
                 speaker = speakers[i]
@@ -757,35 +896,32 @@ def generate_app(
             raise HTTPException(status_code=404, detail="該当する話者が見つかりません")
 
         try:
-            policy = (root_dir / f"speaker_info/{speaker_uuid}/policy.md").read_text(
-                "utf-8"
-            )
-            portrait = b64encode_str(
-                (root_dir / f"speaker_info/{speaker_uuid}/portrait.png").read_bytes()
-            )
+            speaker_path = root_dir / "speaker_info" / speaker_uuid
+            # 話者情報の取得
+            # speaker policy
+            policy_path = speaker_path / "policy.md"
+            policy = policy_path.read_text("utf-8")
+            # speaker portrait
+            portrait_path = speaker_path / "portrait.png"
+            portrait = b64encode_str(portrait_path.read_bytes())
+            # スタイル情報の取得
             style_infos = []
             for style in speaker["styles"]:
                 id = style["id"]
-                icon = b64encode_str(
-                    (
-                        root_dir / f"speaker_info/{speaker_uuid}/icons/{id}.png"
-                    ).read_bytes()
-                )
-                style_portrait_path = (
-                    root_dir / f"speaker_info/{speaker_uuid}/portraits/{id}.png"
-                )
-                style_portrait = (
-                    b64encode_str(style_portrait_path.read_bytes())
-                    if style_portrait_path.exists()
-                    else None
-                )
+                # style icon
+                style_icon_path = speaker_path / "icons" / f"{id}.png"
+                icon = b64encode_str(style_icon_path.read_bytes())
+                # style portrait
+                style_portrait_path = speaker_path / "portraits" / f"{id}.png"
+                style_portrait = None
+                if style_portrait_path.exists():
+                    style_portrait = b64encode_str(style_portrait_path.read_bytes())
+                # voice samples
                 voice_samples = [
                     b64encode_str(
                         (
-                            root_dir
-                            / "speaker_info/{}/voice_samples/{}_{}.wav".format(
-                                speaker_uuid, id, str(j + 1).zfill(3)
-                            )
+                            speaker_path
+                            / "voice_samples/{}_{}.wav".format(id, str(j + 1).zfill(3))
                         ).read_bytes()
                     )
                     for j in range(3)
@@ -805,118 +941,164 @@ def generate_app(
             raise HTTPException(status_code=500, detail="追加情報が見つかりませんでした")
 
         ret_data = {"policy": policy, "portrait": portrait, "style_infos": style_infos}
+
         return ret_data
 
-    @app.get(
-        "/downloadable_libraries",
-        response_model=List[DownloadableLibrary],
-        tags=["音声ライブラリ管理"],
-    )
-    def downloadable_libraries():
-        """
-        ダウンロード可能な音声ライブラリの情報を返します。
+    if engine_manifest_data.supported_features.manage_library:
 
-        Returns
-        -------
-        ret_data: List[DownloadableLibrary]
-        """
-        if not engine_manifest_data.supported_features.manage_library:
-            raise HTTPException(status_code=404, detail="この機能は実装されていません")
-        return library_manager.downloadable_libraries()
-
-    @app.get(
-        "/installed_libraries",
-        response_model=Dict[str, InstalledLibrary],
-        tags=["音声ライブラリ管理"],
-    )
-    def installed_libraries():
-        """
-        インストールした音声ライブラリの情報を返します。
-
-        Returns
-        -------
-        ret_data: List[DownloadableLibrary]
-        """
-        if not engine_manifest_data.supported_features.manage_library:
-            raise HTTPException(status_code=404, detail="この機能は実装されていません")
-        return library_manager.installed_libraries()
-
-    @app.post(
-        "/install_library/{library_uuid}",
-        status_code=204,
-        tags=["音声ライブラリ管理"],
-    )
-    async def install_library(library_uuid: str, request: Request):
-        """
-        音声ライブラリをインストールします。
-        音声ライブラリのZIPファイルをリクエストボディとして送信してください。
-
-        Parameters
-        ----------
-        library_uuid: str
-            音声ライブラリのID
-        """
-        if not engine_manifest_data.supported_features.manage_library:
-            raise HTTPException(status_code=404, detail="この機能は実装されていません")
-        archive = BytesIO(await request.body())
-        loop = asyncio.get_event_loop()
-        await loop.run_in_executor(
-            None, library_manager.install_library, library_uuid, archive
+        @app.get(
+            "/downloadable_libraries",
+            response_model=list[DownloadableLibraryInfo],
+            tags=["音声ライブラリ管理"],
         )
+        def downloadable_libraries() -> list[DownloadableLibraryInfo]:
+            """
+            ダウンロード可能な音声ライブラリの情報を返します。
+
+            Returns
+            -------
+            ret_data: list[DownloadableLibrary]
+            """
+            if not engine_manifest_data.supported_features.manage_library:
+                raise HTTPException(status_code=404, detail="この機能は実装されていません")
+            return library_manager.downloadable_libraries()
+
+        @app.get(
+            "/installed_libraries",
+            response_model=dict[str, InstalledLibraryInfo],
+            tags=["音声ライブラリ管理"],
+        )
+        def installed_libraries() -> dict[str, InstalledLibraryInfo]:
+            """
+            インストールした音声ライブラリの情報を返します。
+
+            Returns
+            -------
+            ret_data: dict[str, InstalledLibrary]
+            """
+            if not engine_manifest_data.supported_features.manage_library:
+                raise HTTPException(status_code=404, detail="この機能は実装されていません")
+            return library_manager.installed_libraries()
+
+        @app.post(
+            "/install_library/{library_uuid}",
+            status_code=204,
+            tags=["音声ライブラリ管理"],
+            dependencies=[Depends(check_disabled_mutable_api)],
+        )
+        async def install_library(
+            library_uuid: str,
+            request: Request,
+        ) -> Response:
+            """
+            音声ライブラリをインストールします。
+            音声ライブラリのZIPファイルをリクエストボディとして送信してください。
+
+            Parameters
+            ----------
+            library_uuid: str
+                音声ライブラリのID
+            """
+            if not engine_manifest_data.supported_features.manage_library:
+                raise HTTPException(status_code=404, detail="この機能は実装されていません")
+            archive = BytesIO(await request.body())
+            loop = asyncio.get_event_loop()
+            await loop.run_in_executor(
+                None, library_manager.install_library, library_uuid, archive
+            )
+            return Response(status_code=204)
+
+        @app.post(
+            "/uninstall_library/{library_uuid}",
+            status_code=204,
+            tags=["音声ライブラリ管理"],
+            dependencies=[Depends(check_disabled_mutable_api)],
+        )
+        def uninstall_library(library_uuid: str) -> Response:
+            """
+            音声ライブラリをアンインストールします。
+
+            Parameters
+            ----------
+            library_uuid: str
+                音声ライブラリのID
+            """
+            if not engine_manifest_data.supported_features.manage_library:
+                raise HTTPException(status_code=404, detail="この機能は実装されていません")
+            library_manager.uninstall_library(library_uuid)
+            return Response(status_code=204)
+
+    @app.post("/initialize_style_id", status_code=204, tags=["その他"])
+    def initialize_style_id(
+        style_id: StyleId,
+        skip_reinit: bool = Query(  # noqa: B008
+            False, description="既に初期化済みのスタイルの再初期化をスキップするかどうか"
+        ),
+        core_version: str | None = None,
+    ) -> Response:
+        """
+        指定されたstyle_idのスタイルを初期化します。
+        実行しなくても他のAPIは使用できますが、初回実行時に時間がかかることがあります。
+        """
+        core = get_core(core_version)
+        core.initialize_style_id_synthesis(style_id, skip_reinit=skip_reinit)
         return Response(status_code=204)
 
-    @app.post(
-        "/uninstall_library/{library_uuid}",
-        status_code=204,
-        tags=["音声ライブラリ管理"],
-    )
-    def uninstall_library(library_uuid: str):
+    @app.get("/is_initialized_style_id", response_model=bool, tags=["その他"])
+    def is_initialized_style_id(
+        style_id: StyleId,
+        core_version: str | None = None,
+    ) -> bool:
         """
-        音声ライブラリをアンインストールします。
-
-        Parameters
-        ----------
-        library_uuid: str
-            音声ライブラリのID
+        指定されたstyle_idのスタイルが初期化されているかどうかを返します。
         """
-        if not engine_manifest_data.supported_features.manage_library:
-            raise HTTPException(status_code=404, detail="この機能は実装されていません")
-        library_manager.uninstall_library(library_uuid)
-        return Response(status_code=204)
+        core = get_core(core_version)
+        return core.is_initialized_style_id_synthesis(style_id)
 
-    @app.post("/initialize_speaker", status_code=204, tags=["その他"])
+    @app.post("/initialize_speaker", status_code=204, tags=["その他"], deprecated=True)
     def initialize_speaker(
-        speaker: int,
+        speaker: StyleId,
         skip_reinit: bool = Query(  # noqa: B008
             False, description="既に初期化済みの話者の再初期化をスキップするかどうか"
         ),
-        core_version: Optional[str] = None,
-    ):
+        core_version: str | None = None,
+    ) -> Response:
         """
-        指定されたspeaker_idの話者を初期化します。
-        実行しなくても他のAPIは使用できますが、初回実行時に時間がかかることがあります。
+        こちらのAPIは非推奨です。`initialize_style_id`を利用してください。
         """
-        engine = get_engine(core_version)
-        engine.initialize_speaker_synthesis(speaker_id=speaker, skip_reinit=skip_reinit)
-        return Response(status_code=204)
+        warnings.warn(
+            "使用しているAPI(/initialize_speaker)は非推奨です。/initialized_style_idを利用してください。",
+            stacklevel=1,
+        )
+        return initialize_style_id(
+            speaker, skip_reinit=skip_reinit, core_version=core_version
+        )
 
-    @app.get("/is_initialized_speaker", response_model=bool, tags=["その他"])
-    def is_initialized_speaker(speaker: int, core_version: Optional[str] = None):
+    @app.get(
+        "/is_initialized_speaker", response_model=bool, tags=["その他"], deprecated=True
+    )
+    def is_initialized_speaker(
+        speaker: StyleId,
+        core_version: str | None = None,
+    ) -> bool:
         """
-        指定されたspeaker_idの話者が初期化されているかどうかを返します。
+        こちらのAPIは非推奨です。`is_initialize_style_id`を利用してください。
         """
-        engine = get_engine(core_version)
-        return engine.is_initialized_speaker_synthesis(speaker)
+        warnings.warn(
+            "使用しているAPI(/is_initialize_speaker)は非推奨です。/is_initialized_style_idを利用してください。",
+            stacklevel=1,
+        )
+        return is_initialized_style_id(speaker, core_version=core_version)
 
-    @app.get("/user_dict", response_model=Dict[str, UserDictWord], tags=["ユーザー辞書"])
-    def get_user_dict_words():
+    @app.get("/user_dict", response_model=dict[str, UserDictWord], tags=["ユーザー辞書"])
+    def get_user_dict_words() -> dict[str, UserDictWord]:
         """
         ユーザー辞書に登録されている単語の一覧を返します。
         単語の表層形(surface)は正規化済みの物を返します。
 
         Returns
         -------
-        Dict[str, UserDictWord]
+        dict[str, UserDictWord]
             単語のUUIDとその詳細
         """
         try:
@@ -925,14 +1107,19 @@ def generate_app(
             traceback.print_exc()
             raise HTTPException(status_code=422, detail="辞書の読み込みに失敗しました。")
 
-    @app.post("/user_dict_word", response_model=str, tags=["ユーザー辞書"])
+    @app.post(
+        "/user_dict_word",
+        response_model=str,
+        tags=["ユーザー辞書"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
     def add_user_dict_word(
         surface: str,
         pronunciation: str,
         accent_type: int,
-        word_type: Optional[WordTypes] = None,
-        priority: Optional[conint(ge=MIN_PRIORITY, le=MAX_PRIORITY)] = None,
-    ):
+        word_type: WordTypes | None = None,
+        priority: Annotated[int | None, Query(ge=MIN_PRIORITY, le=MAX_PRIORITY)] = None,
+    ) -> Response:
         """
         ユーザー辞書に言葉を追加します。
 
@@ -966,15 +1153,20 @@ def generate_app(
             traceback.print_exc()
             raise HTTPException(status_code=422, detail="ユーザー辞書への追加に失敗しました。")
 
-    @app.put("/user_dict_word/{word_uuid}", status_code=204, tags=["ユーザー辞書"])
+    @app.put(
+        "/user_dict_word/{word_uuid}",
+        status_code=204,
+        tags=["ユーザー辞書"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
     def rewrite_user_dict_word(
         surface: str,
         pronunciation: str,
         accent_type: int,
         word_uuid: str,
-        word_type: Optional[WordTypes] = None,
-        priority: Optional[conint(ge=MIN_PRIORITY, le=MAX_PRIORITY)] = None,
-    ):
+        word_type: WordTypes | None = None,
+        priority: Annotated[int | None, Query(ge=MIN_PRIORITY, le=MAX_PRIORITY)] = None,
+    ) -> Response:
         """
         ユーザー辞書に登録されている言葉を更新します。
 
@@ -1013,8 +1205,13 @@ def generate_app(
             traceback.print_exc()
             raise HTTPException(status_code=422, detail="ユーザー辞書の更新に失敗しました。")
 
-    @app.delete("/user_dict_word/{word_uuid}", status_code=204, tags=["ユーザー辞書"])
-    def delete_user_dict_word(word_uuid: str):
+    @app.delete(
+        "/user_dict_word/{word_uuid}",
+        status_code=204,
+        tags=["ユーザー辞書"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
+    def delete_user_dict_word(word_uuid: str) -> Response:
         """
         ユーザー辞書に登録されている言葉を削除します。
 
@@ -1032,16 +1229,22 @@ def generate_app(
             traceback.print_exc()
             raise HTTPException(status_code=422, detail="ユーザー辞書の更新に失敗しました。")
 
-    @app.post("/import_user_dict", status_code=204, tags=["ユーザー辞書"])
+    @app.post(
+        "/import_user_dict",
+        status_code=204,
+        tags=["ユーザー辞書"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
     def import_user_dict_words(
-        import_dict_data: Dict[str, UserDictWord], override: bool
-    ):
+        import_dict_data: dict[str, UserDictWord],
+        override: bool,
+    ) -> Response:
         """
         他のユーザー辞書をインポートします。
 
         Parameters
         ----------
-        import_dict_data: Dict[str, UserDictWord]
+        import_dict_data: dict[str, UserDictWord]
             インポートするユーザー辞書のデータ
         override: bool
             重複したエントリがあった場合、上書きするかどうか
@@ -1055,9 +1258,9 @@ def generate_app(
 
     @app.get("/supported_devices", response_model=SupportedDevicesInfo, tags=["その他"])
     def supported_devices(
-        core_version: Optional[str] = None,
-    ):
-        supported_devices = get_engine(core_version).supported_devices
+        core_version: str | None = None,
+    ) -> Response:
+        supported_devices = get_core(core_version).supported_devices
         if supported_devices is None:
             raise HTTPException(status_code=422, detail="非対応の機能です。")
         return Response(
@@ -1066,14 +1269,14 @@ def generate_app(
         )
 
     @app.get("/engine_manifest", response_model=EngineManifest, tags=["その他"])
-    def engine_manifest():
+    def engine_manifest() -> EngineManifest:
         return engine_manifest_data
 
     @app.post(
         "/validate_kana",
         response_model=bool,
         tags=["その他"],
-        summary="テキストがAquesTalkライクな記法に従っているか判定する",
+        summary="テキストがAquesTalk 風記法に従っているか判定する",
         responses={
             400: {
                 "description": "テキストが不正です",
@@ -1081,9 +1284,9 @@ def generate_app(
             }
         },
     )
-    def validate_kana(text: str):
+    def validate_kana(text: str) -> bool:
         """
-        テキストがAquesTalkライクな記法に従っているかどうかを判定します。
+        テキストがAquesTalk 風記法に従っているかどうかを判定します。
         従っていない場合はエラーが返ります。
 
         Parameters
@@ -1100,10 +1303,14 @@ def generate_app(
                 detail=ParseKanaBadRequest(err).dict(),
             )
 
-    @app.get("/setting", response_class=HTMLResponse, tags=["設定"])
-    def setting_get(request: Request):
+    @app.get("/setting", response_class=Response, tags=["設定"])
+    def setting_get(request: Request) -> Response:
+        """
+        設定ページを返します。
+        """
         settings = setting_loader.load_setting_file()
 
+        brand_name = engine_manifest_data.brand_name
         cors_policy_mode = settings.cors_policy_mode
         allow_origin = settings.allow_origin
 
@@ -1114,17 +1321,25 @@ def generate_app(
             "ui.html",
             {
                 "request": request,
+                "brand_name": brand_name,
                 "cors_policy_mode": cors_policy_mode,
                 "allow_origin": allow_origin,
             },
         )
 
-    @app.post("/setting", response_class=HTMLResponse, tags=["設定"])
+    @app.post(
+        "/setting",
+        response_class=Response,
+        tags=["設定"],
+        dependencies=[Depends(check_disabled_mutable_api)],
+    )
     def setting_post(
-        request: Request,
-        cors_policy_mode: Optional[str] = Form(None),  # noqa: B008
-        allow_origin: Optional[str] = Form(None),  # noqa: B008
-    ):
+        cors_policy_mode: CorsPolicyMode = Form(),  # noqa
+        allow_origin: str | None = Form(default=None),  # noqa
+    ) -> Response:
+        """
+        設定を更新します。
+        """
         settings = Setting(
             cors_policy_mode=cors_policy_mode,
             allow_origin=allow_origin,
@@ -1133,19 +1348,9 @@ def generate_app(
         # 更新した設定へ上書き
         setting_loader.dump_setting_file(settings)
 
-        if allow_origin is None:
-            allow_origin = ""
+        return Response(status_code=204)
 
-        return setting_ui_template.TemplateResponse(
-            "ui.html",
-            {
-                "request": request,
-                "cors_policy_mode": cors_policy_mode,
-                "allow_origin": allow_origin,
-            },
-        )
-
-    # VvlibManifestモデルはAPIとして表には出ないが、エディタ側で利用したいので、手動で追加する
+    # BaseLibraryInfo/VvlibManifestモデルはAPIとして表には出ないが、エディタ側で利用したいので、手動で追加する
     # ref: https://fastapi.tiangolo.com/advanced/extending-openapi/#modify-the-openapi-schema
     def custom_openapi():
         if app.openapi_schema:
@@ -1164,36 +1369,34 @@ def generate_app(
         openapi_schema["components"]["schemas"][
             "VvlibManifest"
         ] = VvlibManifest.schema()
+        # ref_templateを指定しない場合、definitionsを参照してしまうので、手動で指定する
+        base_library_info = BaseLibraryInfo.schema(
+            ref_template="#/components/schemas/{model}"
+        )
+        # definitionsは既存のモデルを重複して定義するため、不要なので削除
+        del base_library_info["definitions"]
+        openapi_schema["components"]["schemas"]["BaseLibraryInfo"] = base_library_info
         app.openapi_schema = openapi_schema
         return openapi_schema
 
-    app.openapi = custom_openapi
+    app.openapi = custom_openapi  # type: ignore[method-assign]
 
     return app
 
 
-if __name__ == "__main__":
+def main() -> None:
     multiprocessing.freeze_support()
 
-    output_log_utf8 = os.getenv("VV_OUTPUT_LOG_UTF8", default="")
-    if output_log_utf8 == "1":
+    output_log_utf8 = decide_boolean_from_env("VV_OUTPUT_LOG_UTF8")
+    if output_log_utf8:
         set_output_log_utf8()
-    elif not (output_log_utf8 == "" or output_log_utf8 == "0"):
-        print(
-            "WARNING:  invalid VV_OUTPUT_LOG_UTF8 environment variable value",
-            file=sys.stderr,
-        )
-
-    default_cors_policy_mode = CorsPolicyMode.localapps
 
     parser = argparse.ArgumentParser(description="VOICEVOX のエンジンです。")
     parser.add_argument(
         "--host", type=str, default="127.0.0.1", help="接続を受け付けるホストアドレスです。"
     )
     parser.add_argument("--port", type=int, default=50021, help="接続を受け付けるポート番号です。")
-    parser.add_argument(
-        "--use_gpu", action="store_true", help="指定するとGPUを使って音声合成するようになります。"
-    )
+    parser.add_argument("--use_gpu", action="store_true", help="GPUを使って音声合成するようになります。")
     parser.add_argument(
         "--voicevox_dir", type=Path, default=None, help="VOICEVOXのディレクトリパスです。"
     )
@@ -1214,12 +1417,12 @@ if __name__ == "__main__":
     parser.add_argument(
         "--enable_mock",
         action="store_true",
-        help="指定するとVOICEVOX COREを使わずモックで音声合成を行います。",
+        help="VOICEVOX COREを使わずモックで音声合成を行います。",
     )
     parser.add_argument(
         "--enable_cancellable_synthesis",
         action="store_true",
-        help="指定すると音声合成を途中でキャンセルできるようになります。",
+        help="音声合成を途中でキャンセルできるようになります。",
     )
     parser.add_argument(
         "--init_processes",
@@ -1228,7 +1431,7 @@ if __name__ == "__main__":
         help="cancellable_synthesis機能の初期化時に生成するプロセス数です。",
     )
     parser.add_argument(
-        "--load_all_models", action="store_true", help="指定すると起動時に全ての音声合成モデルを読み込みます。"
+        "--load_all_models", action="store_true", help="起動時に全ての音声合成モデルを読み込みます。"
     )
 
     # 引数へcpu_num_threadsの指定がなければ、環境変数をロールします。
@@ -1239,8 +1442,8 @@ if __name__ == "__main__":
         type=int,
         default=os.getenv("VV_CPU_NUM_THREADS") or None,
         help=(
-            "音声合成を行うスレッド数です。指定しないと、代わりに環境変数VV_CPU_NUM_THREADSの値が使われます。"
-            "VV_CPU_NUM_THREADSが空文字列でなく数値でもない場合はエラー終了します。"
+            "音声合成を行うスレッド数です。指定しない場合、代わりに環境変数 VV_CPU_NUM_THREADS の値が使われます。"
+            "VV_CPU_NUM_THREADS が空文字列でなく数値でもない場合はエラー終了します。"
         ),
     )
 
@@ -1248,7 +1451,7 @@ if __name__ == "__main__":
         "--output_log_utf8",
         action="store_true",
         help=(
-            "指定するとログ出力をUTF-8でおこないます。指定しないと、代わりに環境変数 VV_OUTPUT_LOG_UTF8 の値が使われます。"
+            "ログ出力をUTF-8でおこないます。指定しない場合、代わりに環境変数 VV_OUTPUT_LOG_UTF8 の値が使われます。"
             "VV_OUTPUT_LOG_UTF8 の値が1の場合はUTF-8で、0または空文字、値がない場合は環境によって自動的に決定されます。"
         ),
     )
@@ -1262,15 +1465,42 @@ if __name__ == "__main__":
             "CORSの許可モード。allまたはlocalappsが指定できます。allはすべてを許可します。"
             "localappsはオリジン間リソース共有ポリシーを、app://.とlocalhost関連に限定します。"
             "その他のオリジンはallow_originオプションで追加できます。デフォルトはlocalapps。"
+            "このオプションは--setting_fileで指定される設定ファイルよりも優先されます。"
         ),
     )
 
     parser.add_argument(
-        "--allow_origin", nargs="*", help="許可するオリジンを指定します。スペースで区切ることで複数指定できます。"
+        "--allow_origin",
+        nargs="*",
+        help=(
+            "許可するオリジンを指定します。スペースで区切ることで複数指定できます。"
+            "このオプションは--setting_fileで指定される設定ファイルよりも優先されます。"
+        ),
     )
 
     parser.add_argument(
         "--setting_file", type=Path, default=USER_SETTING_PATH, help="設定ファイルを指定できます。"
+    )
+
+    parser.add_argument(
+        "--preset_file",
+        type=Path,
+        default=None,
+        help=(
+            "プリセットファイルを指定できます。"
+            "指定がない場合、環境変数 VV_PRESET_FILE、--voicevox_dirのpresets.yaml、"
+            "実行ファイルのディレクトリのpresets.yamlを順に探します。"
+        ),
+    )
+
+    parser.add_argument(
+        "--disable_mutable_api",
+        action="store_true",
+        help=(
+            "辞書登録や設定変更など、エンジンの静的なデータを変更するAPIを無効化します。"
+            "指定しない場合、代わりに環境変数 VV_DISABLE_MUTABLE_API の値が使われます。"
+            "VV_DISABLE_MUTABLE_API の値が1の場合は無効化で、0または空文字、値がない場合は無視されます。"
+        ),
     )
 
     args = parser.parse_args()
@@ -1278,35 +1508,55 @@ if __name__ == "__main__":
     if args.output_log_utf8:
         set_output_log_utf8()
 
-    cpu_num_threads: Optional[int] = args.cpu_num_threads
+    # Synthesis Engine
+    use_gpu: bool = args.use_gpu
+    voicevox_dir: Path | None = args.voicevox_dir
+    voicelib_dirs: list[Path] | None = args.voicelib_dir
+    runtime_dirs: list[Path] | None = args.runtime_dir
+    enable_mock: bool = args.enable_mock
+    cpu_num_threads: int | None = args.cpu_num_threads
+    load_all_models: bool = args.load_all_models
 
-    synthesis_engines = make_synthesis_engines(
-        use_gpu=args.use_gpu,
-        voicelib_dirs=args.voicelib_dir,
-        voicevox_dir=args.voicevox_dir,
-        runtime_dirs=args.runtime_dir,
+    cores = initialize_cores(
+        use_gpu=use_gpu,
+        voicelib_dirs=voicelib_dirs,
+        voicevox_dir=voicevox_dir,
+        runtime_dirs=runtime_dirs,
         cpu_num_threads=cpu_num_threads,
-        enable_mock=args.enable_mock,
-        load_all_models=args.load_all_models,
+        enable_mock=enable_mock,
+        load_all_models=load_all_models,
     )
-    assert len(synthesis_engines) != 0, "音声合成エンジンがありません。"
-    latest_core_version = get_latest_core_version(versions=synthesis_engines.keys())
+    tts_engines = make_tts_engines_from_cores(cores)
+    assert len(tts_engines) != 0, "音声合成エンジンがありません。"
+    latest_core_version = get_latest_core_version(versions=list(tts_engines.keys()))
 
-    cancellable_engine = None
-    if args.enable_cancellable_synthesis:
-        cancellable_engine = CancellableEngine(args)
+    # Cancellable Engine
+    enable_cancellable_synthesis: bool = args.enable_cancellable_synthesis
+    init_processes: int = args.init_processes
 
-    root_dir = args.voicevox_dir if args.voicevox_dir is not None else engine_root()
+    cancellable_engine: CancellableEngine | None = None
+    if enable_cancellable_synthesis:
+        cancellable_engine = CancellableEngine(
+            init_processes=init_processes,
+            use_gpu=use_gpu,
+            voicelib_dirs=voicelib_dirs,
+            voicevox_dir=voicevox_dir,
+            runtime_dirs=runtime_dirs,
+            cpu_num_threads=cpu_num_threads,
+            enable_mock=enable_mock,
+        )
+
+    root_dir: Path | None = voicevox_dir
+    if root_dir is None:
+        root_dir = engine_root()
 
     setting_loader = SettingLoader(args.setting_file)
 
     settings = setting_loader.load_setting_file()
 
-    cors_policy_mode = (
-        args.cors_policy_mode
-        if args.cors_policy_mode is not None
-        else settings.cors_policy_mode
-    )
+    cors_policy_mode: CorsPolicyMode | None = args.cors_policy_mode
+    if cors_policy_mode is None:
+        cors_policy_mode = settings.cors_policy_mode
 
     allow_origin = None
     if args.allow_origin is not None:
@@ -1314,15 +1564,45 @@ if __name__ == "__main__":
     elif settings.allow_origin is not None:
         allow_origin = settings.allow_origin.split(" ")
 
+    # Preset Manager
+    # preset_pathの優先順: 引数、環境変数、voicevox_dir、実行ファイルのディレクトリ
+    # ファイルの存在に関わらず、優先順で最初に指定されたパスをプリセットファイルとして使用する
+    preset_path: Path | None = args.preset_file
+    if preset_path is None:
+        # 引数 --preset_file の指定がない場合
+        env_preset_path = os.getenv("VV_PRESET_FILE")
+        if env_preset_path is not None and len(env_preset_path) != 0:
+            # 環境変数 VV_PRESET_FILE の指定がある場合
+            preset_path = Path(env_preset_path)
+        else:
+            # 環境変数 VV_PRESET_FILE の指定がない場合
+            preset_path = root_dir / "presets.yaml"
+
+    preset_manager = PresetManager(
+        preset_path=preset_path,
+    )
+
+    disable_mutable_api: bool = args.disable_mutable_api | decide_boolean_from_env(
+        "VV_DISABLE_MUTABLE_API"
+    )
+
     uvicorn.run(
         generate_app(
-            synthesis_engines,
+            tts_engines,
+            cores,
             latest_core_version,
             setting_loader,
+            preset_manager=preset_manager,
+            cancellable_engine=cancellable_engine,
             root_dir=root_dir,
             cors_policy_mode=cors_policy_mode,
             allow_origin=allow_origin,
+            disable_mutable_api=disable_mutable_api,
         ),
         host=args.host,
         port=args.port,
     )
+
+
+if __name__ == "__main__":
+    main()
