@@ -1,10 +1,10 @@
 import base64
 import io
-from typing import List, Tuple
 
 import numpy as np
 import soundfile
-from scipy.signal import resample
+from numpy.typing import NDArray
+from soxr import resample
 
 
 class ConnectBase64WavesException(Exception):
@@ -12,7 +12,7 @@ class ConnectBase64WavesException(Exception):
         self.message = message
 
 
-def decode_base64_waves(waves: List[str]) -> List[Tuple[np.ndarray, int]]:
+def decode_base64_waves(waves: list[str]) -> list[tuple[NDArray[np.float64], int]]:
     """
     base64エンコードされた複数のwavデータをデコードする
     Parameters
@@ -21,7 +21,7 @@ def decode_base64_waves(waves: List[str]) -> List[Tuple[np.ndarray, int]]:
         base64エンコードされたwavデータのリスト
     Returns
     -------
-    waves_nparray_sr: List[Tuple[np.ndarray, int]]
+    waves_nparray_sr: List[Tuple[NDArray[np.float64], int]]
         (NumPy配列の音声波形データ, サンプリングレート) 形式のタプルのリスト
     """
     if len(waves) == 0:
@@ -42,7 +42,7 @@ def decode_base64_waves(waves: List[str]) -> List[Tuple[np.ndarray, int]]:
     return waves_nparray_sr
 
 
-def connect_base64_waves(waves: List[str]) -> Tuple[np.ndarray, int]:
+def connect_base64_waves(waves: list[str]) -> tuple[NDArray[np.float64], int]:
     waves_nparray_sr = decode_base64_waves(waves)
 
     max_sampling_rate = max([sr for _, sr in waves_nparray_sr])
@@ -52,7 +52,7 @@ def connect_base64_waves(waves: List[str]) -> Tuple[np.ndarray, int]:
     waves_nparray_list = []
     for nparray, sr in waves_nparray_sr:
         if sr != max_sampling_rate:
-            nparray = resample(nparray, max_sampling_rate * len(nparray) // sr)
+            nparray = resample(nparray, sr, max_sampling_rate)
         if nparray.ndim < max_channels:
             nparray = np.array([nparray, nparray]).T
         waves_nparray_list.append(nparray)
