@@ -34,14 +34,14 @@ def test_release_build(dist_dir: Path, skip_run_process: bool) -> None:
     # テキスト -> クエリ
     text = "こんにちは、音声合成の世界へようこそ"
     req = Request(
-        base_url + "audio_query?" + urlencode({"speaker": "1", "text": text}),
+        base_url + "audio_query?" + urlencode({"style_id": "1", "text": text}),
         method="POST",
     )
     with urlopen(req) as res:
         query = json.loads(res.read().decode("utf-8"))
 
     # クエリ -> 音声
-    req = Request(base_url + "synthesis?speaker=1", method="POST")
+    req = Request(base_url + "synthesis?style_id=1", method="POST")
     req.add_header("Content-Type", "application/json")
     req.data = json.dumps(query).encode("utf-8")
     with urlopen(req) as res:
@@ -56,6 +56,7 @@ def test_release_build(dist_dir: Path, skip_run_process: bool) -> None:
 
     if not skip_run_process:
         # プロセスが稼働中であることを確認
+        assert process is not None
         assert process.poll() is None
 
         # 停止
