@@ -2,18 +2,51 @@ import os
 import sys
 import traceback
 from pathlib import Path
+from typing import Literal
 
-from appdirs import user_data_dir
+from platformdirs import user_data_dir
+
+
+def _runtime_type() -> Literal["nuitka", "pyinstaller", "python"]:
+    """
+    コンパイルに使用したライブラリ名を返す。
+    コンパイルしていない場合は"python"を返す。
+    """
+    # nuitkaビルドをした際はグローバルに__compiled__が含まれる
+    if "__compiled__" in globals():
+        return "nuitka"
+
+    # pyinstallerでビルドをした際はsys.frozenが設定される
+    elif getattr(sys, "frozen", False) and hasattr(sys, "_MEIPASS"):
+        return "pyinstaller"
+
+    return "python"
 
 
 def engine_root() -> Path:
-    if is_development():
-        root_dir = Path(__file__).parents[2]
-
-    # Nuitka/Pyinstallerでビルドされている場合
-    else:
+    """
+    開発環境ではリポジトリのルートディレクトリを返す。
+    コンパイル後は実行ファイルがあるディレクトリを返す。
+    """
+    runtime = _runtime_type()
+    if runtime == "nuitka":
         root_dir = Path(sys.argv[0]).parent
 
+    elif runtime == "pyinstaller":
+        root_dir = Path(sys.executable).parent
+
+    else:
+        root_dir = Path(__file__).parents[2]
+
+    return root_dir.resolve(strict=True)
+
+
+def internal_root() -> Path:
+    """
+    コンパイル時に収集された実行ファイル内部用のルートディレクトリを返す。
+    開発環境ではリポジトリのルートディレクトリを返す。
+    """
+    root_dir = Path(__file__).parents[2]
     return root_dir.resolve(strict=True)
 
 
@@ -22,15 +55,7 @@ def is_development() -> bool:
     開発版かどうか判定する関数
     Nuitka/Pyinstallerでコンパイルされていない場合は開発環境とする。
     """
-    # nuitkaビルドをした際はグローバルに__compiled__が含まれる
-    if "__compiled__" in globals():
-        return False
-
-    # pyinstallerでビルドをした際はsys.frozenが設定される
-    elif getattr(sys, "frozen", False):
-        return False
-
-    return True
+    return _runtime_type() == "python"
 
 
 def get_save_dir():
